refactor(init): split into 3 sibling modules — orchestration / resolve / scaffold (sage P148 cycle 7)
Sage cycle 7 Architecture lens: "file has grown to ~650+ lines mixing
three distinct concerns (resolution, content definition for 5 artifact
types, and execution)". Split addresses the concern by giving each
concern its own module.

New layout (flat in src/commands/, matching every other command):

  - src/commands/init.ts (126 lines) — orchestration only:
      * `init()` public entry point
      * `validateTargetDir()` helper extracting the 45-line lstat /
        symlink / ENOENT preamble sage P148 cycle 6 flagged for
        extraction
      * Re-exports public types from the two sibling modules so the
        cli.ts and test imports keep working unchanged

  - src/commands/init-resolve.ts (80 lines) — pure name + targetDir
    resolution:
      * `ResolvedInitTarget` discriminated union
      * `resolveInitTarget()` with `platformOverride` test seam
      * No I/O, no filesystem

  - src/commands/init-scaffold.ts (460 lines) — content definitions:
      * `ScaffoldEntry { path, content }` interface
      * `scaffoldEntriesFor()` dispatcher
      * All per-type builders (manifest, package.json, README, type-
        specific files) and module constants (GITIGNORE,
        SKILL_WORKFLOW_MAIN, PIPELINE_ACTION_JSON, PIPELINE_ACTION_TS)
      * Adding a new artifact type touches this file only

Public surface unchanged — `./init.js` re-exports `ArtifactInitType`,
`ScaffoldEntry`, `scaffoldEntriesFor`, `ResolvedInitTarget`,
`resolveInitTarget`. CLI imports + test imports work without changes.

CLI smoke verified: scaffold output identical to pre-split for all
three #107 reproduction cases. Full suite: 861/0.

Closes the sage cycle 7 file-organization concern. The remaining
recurring concern across cycles 4-7 (parallel-list drift between
`scaffoldFilesFor` and `Bun.write`) was already closed in the
cycle-7 data-driven refactor — this cycle just gives the now-clean
data-driven shape three coherent homes.
